Moving Theme Code into Own File (#54)
* theme code separated into own file

* fixing theme references

* Removing unused imports

* small fix with theme code

* Updating theme implementation

---------
